combine fim and hosts
