--- server/service/service_hosts.go
@@ -1,32 +1,0 @@
-package service
-
-import (
-	"context"
-
-	"github.com/kolide/fleet/server/kolide"
-)
-
-func (svc service) ListHosts(ctx context.Context, opt kolide.ListOptions) ([]*kolide.Host, error) {
-	return svc.ds.ListHosts(opt)
-}
-
-func (svc service) GetHost(ctx context.Context, id uint) (*kolide.Host, error) {
-	return svc.ds.Host(id)
-}
-
-func (svc service) GetHostSummary(ctx context.Context) (*kolide.HostSummary, error) {
-	online, offline, mia, new, err := svc.ds.GenerateHostStatusStatistics(svc.clock.Now())
-	if err != nil {
-		return nil, err
-	}
-	return &kolide.HostSummary{
-		OnlineCount:  online,
-		OfflineCount: offline,
-		MIACount:     mia,
-		NewCount:     new,
-	}, nil
-}
-
-func (svc service) DeleteHost(ctx context.Context, id uint) error {
-	return svc.ds.DeleteHost(id)
-}
